Add first badge, fix styling bugs with flex-shrink
