+++ src/components/NotificationBadgesContainer/NotificationBadgesContainer.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import TestBadge from '../NotificationBadges/TestBadge'
+import AccountCreatedBadge from '../NotificationBadges/AccountCreated'
 import { PlaylistContext } from '../../contexts/playlistContext'
 import { useContext } from 'react'
 import moment from 'moment'
@@ -8,12 +8,12 @@ const Wrapper = styled.div`
   background: #f9f9f9;
   display: flex;
   align-items: center;
+  justify-content: center;
   width: 65vw;
   height: 12.5rem;
   border-radius: ${({ theme }) => theme.styles.borderRadius};
   box-shadow: ${({ theme }) => theme.styles.boxShadowLight};
   align-self: center;
-  display: flex;
   overflow-y: scroll;
 `
 
@@ -23,20 +23,20 @@ const NotificationBadgesContainer = () => {
   console.log(moment().format('MM/DD/YYYY'));
   return (
     <Wrapper>
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
-      <TestBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
+      <AccountCreatedBadge />
     </Wrapper>
   )
 }
